change return type calculateTime

+++ app.go
@@ -82,7 +82,7 @@ func (a *App) UpdateTodoList(id uint, listID uint) error {
 	return todo.UpdateTodoList(id, listID)
 }
 
-func (a *App) CalculateRemainingTime(id uint) (float64, error) {
+func (a *App) CalculateRemainingTime(id uint) (int, error) {
 	return todo.CalculateDaysLeft(id)
 }
